log-consumer/requirements.txt
kafka-python==2.0.2
mysql-connector-python==8.0.26
python-dotenv==0.19.0
requests==2.31.0 